9. ignore feature 를 user 도메인에 병합
