+++ src/main/java/com/nexters/phochak/user/adapter/in/web/UserController.java
@@ -3,6 +3,7 @@ package com.nexters.phochak.user.adapter.in.web;
 import com.nexters.phochak.auth.Auth;
 import com.nexters.phochak.auth.UserContext;
 import com.nexters.phochak.post.application.port.in.CommonResponseDto;
+import com.nexters.phochak.user.application.port.in.IgnoredUserResponseDto;
 import com.nexters.phochak.user.application.port.in.JwtResponseDto;
 import com.nexters.phochak.user.application.port.in.JwtTokenUseCase;
 import com.nexters.phochak.user.application.port.in.LoginRequestDto;
@@ -13,9 +14,11 @@ import com.nexters.phochak.user.application.port.in.UserCheckResponseDto;
 import com.nexters.phochak.user.application.port.in.UserInfoResponseDto;
 import com.nexters.phochak.user.application.port.in.UserUseCase;
 import com.nexters.phochak.user.application.port.in.WithdrawRequestDto;
+import com.nexters.phochak.user.application.port.out.IgnoredUserUseCase;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -25,12 +28,15 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.util.List;
+
 @Slf4j
 @RequiredArgsConstructor
 @RequestMapping("/v1/user")
 @RestController
 public class UserController {
     private final UserUseCase userUseCase;
+    private final IgnoredUserUseCase ignoredUserUseCase;
     private final JwtTokenUseCase jwtTokenUseCase;
 
     @GetMapping("/login/{provider}")
@@ -84,5 +90,27 @@ public class UserController {
         return new CommonResponseDto<>(userUseCase.getInfo(userId, pageOwnerId));
     }
 
+    @Auth
+    @GetMapping("/ignore")
+    public CommonResponseDto<List<IgnoredUserResponseDto>> getIgnoreUser() {
+        Long me = UserContext.CONTEXT.get();
+        return new CommonResponseDto<>(ignoredUserUseCase.getIgnoreUserList(me));
+    }
+
+    @Auth
+    @PostMapping("/ignore/{ignoredUserId}")
+    public CommonResponseDto<Void> ignoreUser(@PathVariable(value = "ignoredUserId") Long ignoredUserId) {
+        Long me = UserContext.CONTEXT.get();
+        ignoredUserUseCase.ignoreUser(me, ignoredUserId);
+        return new CommonResponseDto<>();
+    }
+
+    @Auth
+    @DeleteMapping("/ignore/{ignoredUserId}")
+    public CommonResponseDto<Void> cancelIgnoreUser(@PathVariable(value = "ignoredUserId") Long ignoredUserId) {
+        Long me = UserContext.CONTEXT.get();
+        ignoredUserUseCase.cancelIgnoreUser(me, ignoredUserId);
+        return new CommonResponseDto<>();
+    }
 
 }
