+++ src/main/java/com/nexters/phochak/user/application/port/in/IgnoredUserResponseDto.java
@@ -1,6 +1,6 @@
-package com.nexters.phochak.ignore.application.port.in;
+package com.nexters.phochak.user.application.port.in;
 
-import com.nexters.phochak.ignore.domain.IgnoredUser;
+import com.nexters.phochak.user.domain.IgnoredUser;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
