+++ src/test/java/com/nexters/phochak/ignore/IgnoreControllerTest.java
@@ -6,9 +6,9 @@ import com.nexters.phochak.common.RestDocsApiTest;
 import com.nexters.phochak.common.Scenario;
 import com.nexters.phochak.common.TestUtil;
 import com.nexters.phochak.ignore.adapter.in.web.IgnoreUserController;
-import com.nexters.phochak.ignore.adapter.out.persistence.IgnoredUserEntityRelation;
-import com.nexters.phochak.ignore.adapter.out.persistence.IgnoredUserRepository;
 import com.nexters.phochak.user.adapter.out.api.KakaoInformationFeignClient;
+import com.nexters.phochak.user.adapter.out.persistence.IgnoredUserEntityRelation;
+import com.nexters.phochak.user.adapter.out.persistence.IgnoredUserRepository;
 import com.nexters.phochak.user.adapter.out.persistence.UserEntity;
 import com.nexters.phochak.user.adapter.out.persistence.UserRepository;
 import org.junit.jupiter.api.BeforeEach;
